+++ src/main/java/com/portfolio/portfolio/repositories/PersonaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonaRepository extends JpaRepository <PersonaModel, Long> {
+public interface PersonaRepository extends JpaRepository <PersonaModel, Integer> {
 
 
 }
